lnwallet: move updateLog to its own file.
